+++ lib/acts_cruddy/formats/html.rb
@@ -12,20 +12,8 @@ module ActsCruddy
       def create
 
         if @record.save
-
-          url = case self.class.redirect_to_after_create.to_s
-            when 'edit'
-              url_for(:controller => controller_path, :action => :edit, :id => @record)
-            when 'show'
-              url_for(:controller => controller_path, :action => :show, :id => @record)
-            when 'index'
-              url_for(:controller => controller_path, :action => :index)
-            else
-              url_for(self.class.redirect_to_after_create)
-          end
-
-          redirect_to(url, :notice => hierarchical_translate('record_created', :name => record_name_value(@record)))
-
+          flash[:notice] = hierarchical_translate('record_created', :name => record_name_value(@record))
+          redirect_after_create
         else
           render :action => 'new'
         end
@@ -35,20 +23,8 @@ module ActsCruddy
       def update
         
         if @record.update_attributes(params[record_name])
-
-          url = case self.class.redirect_to_after_update.to_s
-            when 'edit'
-              url_for(:controller => controller_path, :action => :edit, :id => @record)
-            when 'show'
-              url_for(:controller => controller_path, :action => :show, :id => @record)
-            when 'index'
-              url_for(:controller => controller_path, :action => :index)
-            else
-              url_for(self.class.redirect_to_after_update)
-          end
-
-          redirect_to(url, :notice => hierarchical_translate('record_updated', :name => record_name_value(@record)))
-
+          flash[:notice] = hierarchical_translate('record_updated', :name => record_name_value(@record))
+          redirect_after_update
         else
           render :action => 'edit'
         end
@@ -56,15 +32,16 @@ module ActsCruddy
       end
 
       def destroy
-            
-        @record.destroy
 
-        redirect_to(
-          url_for(:controller => controller_path, :action => :index),
-          :notice => hierarchical_translate('record_destroyed', :name => record_name_value(@record)))
+        @record.destroy
+        
+        flash[:notice] = hierarchical_translate('record_destroyed', :name => record_name_value(@record))
+        redirect_after_destroy
 
       end
 
+      protected
+      
       # The scope of a translation by default is based on the location of the
       # template file.  This provides a version of translate that will look
       # for a value based on the application_controller subclass name, then fall back to the
@@ -82,6 +59,25 @@ module ActsCruddy
 
         t(key, options)
 
+      end
+      
+      #
+      # Overwrite the any of the following 3 methods in your acts_cruddy controller if you don't like the defaults
+      #
+      def redirect_after_save
+        redirect_to url_for(:controller => controller_path, :action => :show, :id => @record)        
+      end
+      
+      def redirect_after_create
+        redirect_after_save
+      end
+      
+      def redirect_after_update
+        redirect_after_save
+      end
+      
+      def redirect_after_destroy
+        redirect_to url_for(:controller => controller_path, :action => :index)
       end
 
     end